Remove old lab
