+++ app-frontend/src/app/components/tools/diagramContainer/diagramContainer.component.js
@@ -6,6 +6,7 @@ export default {
     bindings: {
         onPreview: '&',
         onShare: '&',
+        onParameterChange: '&',
         toolDefinition: '<',
         toolParameters: '<'
     }
