--- app-frontend/src/app/components/tools/diagramContainer2/diagramContainer.component.js
@@ -1,13 +1,0 @@
-import diagramContainerTpl from './diagramContainer.html';
-
-export default {
-    templateUrl: diagramContainerTpl,
-    controller: 'DiagramContainerController2',
-    bindings: {
-        onPreview: '&',
-        onShare: '&',
-        onParameterChange: '&',
-        toolDefinition: '<',
-        toolParameters: '<'
-    }
-};
