+++ app-frontend/src/app/pages/lab/run/navbar/navbar.controller.js
