--- app-frontend/src/app/pages/lab/run2/run.html
@@ -1,61 +1,0 @@
-<!-- Workflow container -->
-<div class="main with-position">
-  <div class="lab-container">
-    <div class="lab-left"
-         ng-show="$ctrl.isShowingPreview && !!$ctrl.showMap"
-         ng-style="$ctrl.labLeftStyle" draggable="false">
-      <rf-lab-map map-id="lab-preview" on-close="$ctrl.onPreviewClose()"></rf-lab-map>
-    </div>
-    <div class="resize-separator"
-         ng-style="$ctrl.resizeHandleStyle"
-         ng-show="$ctrl.isShowingPreview"
-         draggable="false">
-      <div class="resize-handle-grabber"
-           ng-mousedown="$ctrl.startResize($event)"
-           draggable="false"
-           title="Drag to resize Preview"
-      >
-        <div class="icon-gripper" draggable="false"></div>
-      </div>
-    </div>
-    <div class="resize-ghost"
-         ng-style="$ctrl.labResizingStyle"
-         ng-show="$ctrl.labResizing"
-         draggable="false">
-      <div class="resize-ghost-text" ng-show="$ctrl.splitPercentX && $ctrl.splitPercentX > 90">
-        Maximize Map
-      </div>
-      <div class="resize-ghost-text" ng-show="$ctrl.splitPercentX && $ctrl.splitPercentX < 10">
-        Maximize Diagram
-      </div>
-    </div>
-    <div class="lab-right"
-         ng-style="$ctrl.labRightStyle"
-         ng-show="!!$ctrl.showDiagram || !$ctrl.isShowingPreview"
-         draggable="false">
-      <div class="lab-action-controls">
-        <button class="btn btn-light">Add Note</button>
-        <button class="btn btn-secondary" ng-click="$ctrl.createToolRun()">Apply Changes</button>
-      </div>
-      <rf-diagram-container2 ng-if="$ctrl.tool"
-                             class="with-position"
-                             toolrun-valid="!!$ctrl.lastToolRun"
-                             on-preview="$ctrl.showPreview(data)"
-                             on-share="$ctrl.shareNode(data)"
-                             on-parameter-change="$ctrl.onExecutionParametersChange(sourceId, project, band, override)"
-                             tool-definition="$ctrl.tool.definition"
-                             tool-parameters="$ctrl.toolParameters">
-      </rf-diagram-container2>
-    </div>
-    <div class="lab-footer"
-         ng-show="!$ctrl.lastToolRun && !$ctrl.failedToolRun">
-      <span class="icon-warning"></span>
-      You must apply changes after defining inputs.
-    </div>
-    <div class="lab-footer"
-         ng-show="$ctrl.failedToolRun">
-      <span class="icon-warning"></span>
-      There was an error applying your changes. Please verify that all inputs are defined.
-    </div>
-  </div>
-</div>
