--- app-frontend/src/app/pages/lab/run2/run.module.js
@@ -1,9 +1,0 @@
-import angular from 'angular';
-
-import LabRunController from './run.controller.js';
-
-const labRunModule = angular.module('pages.lab.run2', []);
-
-labRunModule.controller('LabRunController2', LabRunController);
-
-export default labRunModule;
