+++ app-frontend/src/assets/styles/sass/_shame.scss
@@ -1380,11 +1380,6 @@ form.inset {
 }
 
 rf-diagram-container {
-    display: flex;
-    flex: 1;
-}
-
-rf-diagram-container2 {
   display: flex;
   flex: 1;
   overflow: hidden;
